Refactor TargetExtractor to separate class
